+++ cli/src/commands/commit.rs
@@ -154,6 +154,7 @@ new working-copy commit.
     };
     let tree = diff_selector.select(
         Diff::new(&base_tree, &commit.tree()),
+        Diff::new(commit.parents_conflict_label()?, commit.conflict_label()),
         matcher.as_ref(),
         format_instructions,
     )?;