+++ cli/src/commands/split.rs
@@ -554,6 +554,10 @@ The changes that are not selected will replace the original commit.
     let parent_tree = target_commit.parent_tree(tx.repo())?;
     let selected_tree = diff_selector.select(
         Diff::new(&parent_tree, &target_commit.tree()),
+        Diff::new(
+            target_commit.parents_conflict_label()?,
+            target_commit.conflict_label(),
+        ),
         matcher,
         format_instructions,
     )?;